pyproject.toml
[project]
name = "snowflake-fastapi-mcp"
version = "0.1.0"
description = "A FastAPI server that provides Snowflake database access via MCP"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi (>=0.110.0)",
    "uvicorn (==0.23.2)",
    "pydantic (>=2.7.2,<3.0.0)",
    "pydantic-settings (>=2.0.3)",
    "snowflake-snowpark-python (>=1.30.0,<2.0.0)",
    "snowflake-connector-python (>=3.14.0)",
    "python-dotenv (==1.0.0)",
    "pyyaml (==6.0.1)",
    "fastapi-mcp (==0.3.3)",
    "httpx (>=0.25.0)"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
